+++ examples/index.php
@@ -14,7 +14,7 @@
 $loader = require realpath(__DIR__ . '/../vendor/') . DIRECTORY_SEPARATOR . 'autoload.php';
 $loader->addPsr4('Contoh\\', realpath(__DIR__ . '/Demo/'));  // Example of namespace in deep folder
 
-$gubug = new Gubug\Framework();
+$gubug = new Gubug\Framework('dev');
 
 
 // =========== Namespace prefix